+++ cloudmanager/cvo_aws.go
@@ -48,19 +48,20 @@ type createCVOAWSDetails struct {
 
 // haParamsAWS the input for requesting a CVO
 type haParamsAWS struct {
-	ClusterFloatingIP         string   `structs:"clusterFloatingIP,omitempty"`
-	DataFloatingIP            string   `structs:"dataFloatingIP,omitempty"`
-	DataFloatingIP2           string   `structs:"dataFloatingIP2,omitempty"`
-	SvmFloatingIP             string   `structs:"svmFloatingIP,omitempty"`
-	FailoverMode              string   `structs:"failoverMode,omitempty"`
-	Node1SubnetID             string   `structs:"node1SubnetId,omitempty"`
-	Node2SubnetID             string   `structs:"node2SubnetId,omitempty"`
-	MediatorSubnetID          string   `structs:"mediatorSubnetId,omitempty"`
-	MediatorKeyPairName       string   `structs:"mediatorKeyPairName,omitempty"`
-	PlatformSerialNumberNode1 string   `structs:"platformSerialNumberNode1,omitempty"`
-	PlatformSerialNumberNode2 string   `structs:"platformSerialNumberNode2,omitempty"`
-	MediatorAssignPublicIP    *bool    `structs:"mediatorAssignPublicIP,omitempty"`
-	RouteTableIds             []string `structs:"routeTableIds,omitempty"`
+	ClusterFloatingIP           string   `structs:"clusterFloatingIP,omitempty"`
+	DataFloatingIP              string   `structs:"dataFloatingIP,omitempty"`
+	DataFloatingIP2             string   `structs:"dataFloatingIP2,omitempty"`
+	SvmFloatingIP               string   `structs:"svmFloatingIP,omitempty"`
+	FailoverMode                string   `structs:"failoverMode,omitempty"`
+	Node1SubnetID               string   `structs:"node1SubnetId,omitempty"`
+	Node2SubnetID               string   `structs:"node2SubnetId,omitempty"`
+	MediatorSubnetID            string   `structs:"mediatorSubnetId,omitempty"`
+	MediatorKeyPairName         string   `structs:"mediatorKeyPairName,omitempty"`
+	PlatformSerialNumberNode1   string   `structs:"platformSerialNumberNode1,omitempty"`
+	PlatformSerialNumberNode2   string   `structs:"platformSerialNumberNode2,omitempty"`
+	MediatorInstanceProfileName string   `structs:"mediatorInstanceProfileName,omitempty"`
+	MediatorAssignPublicIP      *bool    `structs:"mediatorAssignPublicIP,omitempty"`
+	RouteTableIds               []string `structs:"routeTableIds,omitempty"`
 }
 
 // ebsVolumeSize the input for requesting a CVO
@@ -362,16 +363,19 @@ func validateCVOParams(cvoDetails createCVOAWSDetails) error {
 		return fmt.Errorf("ontap_version parameter not required when having use_latest_version as true")
 	}
 
+	// by Node byol license for existing customers
 	if cvoDetails.VsaMetadata.PlatformSerialNumber != "" && cvoDetails.VsaMetadata.LicenseType != "cot-premium-byol" {
 		return fmt.Errorf("platform_serial_number parameter required only when having license_type as cot-premium-byol")
 	}
 
+	// by Node byol license for existing customers
 	if cvoDetails.IsHA == true && cvoDetails.VsaMetadata.LicenseType == "ha-cot-premium-byol" {
 		if cvoDetails.HAParams.PlatformSerialNumberNode1 == "" || cvoDetails.HAParams.PlatformSerialNumberNode2 == "" {
 			return fmt.Errorf("both platform_serial_number_node1 and platform_serial_number_node2 parameters are required when having ha type as true and license_type as ha-cot-premium-byol")
 		}
 	}
 
+	// by Node byol license for existing customers
 	if cvoDetails.IsHA == false && (cvoDetails.HAParams.PlatformSerialNumberNode1 != "" || cvoDetails.HAParams.PlatformSerialNumberNode2 != "") {
 		return fmt.Errorf("both platform_serial_number_node1 and platform_serial_number_node2 parameters are only required when having ha type as true and license_type as ha-cot-premium-byol")
 	}
@@ -392,6 +396,7 @@ func validateCVOParams(cvoDetails createCVOAWSDetails) error {
 		return fmt.Errorf("subnet_id not required when having ha as true")
 	}
 
+	// by Capacity license
 	if cvoDetails.VsaMetadata.CapacityPackageName != "" {
 		log.Print("Verify cvo parameter capacity_package_name is not empty")
 		if cvoDetails.IsHA == true && cvoDetails.VsaMetadata.LicenseType != "ha-capacity-paygo" {
@@ -403,7 +408,8 @@ func validateCVOParams(cvoDetails createCVOAWSDetails) error {
 	}
 
 	if strings.HasSuffix(cvoDetails.VsaMetadata.LicenseType, "capacity-paygo") && cvoDetails.VsaMetadata.CapacityPackageName == "" {
-		return fmt.Errorf("capacity_package_name is required on selecting Bring Your Own License with capacity based license type or Freemium")
+		return fmt.Errorf("capacity_package_name is required with capacity based license type")
 	}
+
 	return nil
 }